Update isAnchorElement()
Add test for `<a>` with id attribute set. This is necessary to display the anchor actions toolbar for links (linkHref) where the href was created before the anchor (id). In these cases, the 'link' customProperty wins out, and the 'anchor' customProperty isn't set. This is only an issue on initial edit.